# Speedup: only import httpx if needed